requirements.txt
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
playwright>=1.40.0